LFS-18 / LFS-152: Add a service for managing generic permissions Refactor the PermissionsChangeService into an actual service, rather than a static class.
